Refine merge case fixtures
